pyproject.toml
[project]
name = "model-detech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
